+++ R/zz.R
@@ -9,11 +9,11 @@
   registerS3method("semi_join", "reactibble", semi_join.reactibble, asNamespace("dplyr"))
   registerS3method("nest_join", "reactibble", nest_join.reactibble, asNamespace("dplyr"))
   registerS3method("dplyr_reconstruct", "reactibble", dplyr_reconstruct.reactibble, asNamespace("dplyr"))
+  registerS3method("slice", "reactibble", slice.reactibble, asNamespace("dplyr"))
   #registerS3method("as_tibble", "reactibble", as_tibble.reactibble, asNamespace("tibble"))
   registerS3method("vec_ptype_abbr", "reactive_col", vec_ptype_abbr.reactive_col, asNamespace("vctrs"))
   registerS3method("pillar_shaft", "reactive_col", pillar_shaft.reactive_col, asNamespace("pillar"))
   registerS3method("tbl_sum", "reactibble", tbl_sum.reactibble, asNamespace("tibble"))
-
 
   #registerS3method("vec_arith", "default", vec_arith.default, asNamespace("vctrs"))